assert issues fixed for sdle scan

+++ test/python/scripts/tf_unittest_runner.py
@@ -231,7 +231,7 @@ def list_tests(module_list, regex_input):
 
 
 def read_tests_from_file(filename):
-    assert os.path.exists(filename),"Couldn't find the path {}".format(filename)
+    if not os.path.exists(filename): raise AssertionError("Could not find the path {}".format(filename))
     with open(filename) as list_of_tests:
         return [
             line.split('#')[0].rstrip('\n').strip(' ')
@@ -260,7 +260,7 @@ def run_test(test_list, xml_report, verbosity=0):
         for test in test_list:
             names = loader.loadTestsFromName(test)
             suite.addTest(names)
-        assert os.path.exists(xml_report), "Could not find the path"
+        if not os.path.exists(xml_report): raise AssertionError("Could not find the path")
         with open(xml_report, 'wb') as output:
             sys.stdout = open(os.devnull, "w")
             sys.stderr = open(os.devnull, "w")
